+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Added
+  - Support for workflow IDs and attributes (used by Forge Webhooks) in the `Translate Object (Custom)` command.
+
 ## [2.2.0] - 2020-10-23
 
 - Added